refactor: extract client http feature into ferriskey-api-client
