+++ libs/ferriskey-api-client/src/handlers/create_post_logout_redirect_uri.rs
@@ -1,17 +1,13 @@
-use crate::application::http::{
-    client::validators::CreateRedirectUriValidator,
-    server::{
-        api_entities::{
-            api_error::{ApiError, ValidateJson},
-            response::Response,
-        },
-        app_state::AppState,
-    },
-};
+use crate::validators::CreateRedirectUriValidator;
 use axum::{
     Extension,
     extract::{Path, State},
 };
+use ferriskey_api_core::api_entities::{
+    api_error::{ApiError, ValidateJson},
+    response::Response,
+};
+use ferriskey_api_core::app_state::AppState;
 use ferriskey_core::domain::authentication::value_objects::Identity;
 use ferriskey_core::domain::client::entities::redirect_uri::RedirectUri;
 use ferriskey_core::domain::client::value_objects::CreateRedirectUriRequest;
