+++ libs/ferriskey-api-client/src/handlers/delete_client.rs
@@ -1,6 +1,6 @@
-use crate::application::http::server::api_entities::api_error::{ApiError, ApiErrorResponse};
-use crate::application::http::server::api_entities::response::Response;
-use crate::application::http::server::app_state::AppState;
+use ferriskey_api_core::api_entities::api_error::{ApiError, ApiErrorResponse};
+use ferriskey_api_core::api_entities::response::Response;
+use ferriskey_api_core::app_state::AppState;
 
 use axum::{
     Extension,
