+++ libs/ferriskey-api-client/src/handlers/get_client_roles.rs
@@ -1,8 +1,8 @@
-use crate::application::http::server::api_entities::api_error::{ApiError, ApiErrorResponse};
-use crate::application::http::server::api_entities::response::Response;
-use crate::application::http::server::app_state::AppState;
 use axum::Extension;
 use axum::extract::{Path, State};
+use ferriskey_api_core::api_entities::api_error::{ApiError, ApiErrorResponse};
+use ferriskey_api_core::api_entities::response::Response;
+use ferriskey_api_core::app_state::AppState;
 use ferriskey_core::domain::authentication::value_objects::Identity;
 use ferriskey_core::domain::client::entities::GetClientRolesInput;
 use ferriskey_core::domain::client::ports::ClientService;
